refactor: move filter of routes result to another component
